requirements.txt
streamlit
openai-agents
python-dotenv
duckduckgo_search
twilio